+++ Cod3rsGrowth.Infra/ModuloInjetorInfra.cs
@@ -4,7 +4,6 @@ using Cod3rsGrowth.Test.Repositorios;
 using Microsoft.Extensions.DependencyInjection;
 using LinqToDB.AspNet;
 using LinqToDB;
-using Cod3rsGrowth.Infra;
 using FluentMigrator.Runner;
 using DotNetEnv;
