+++ Cod3rsGrowth.Web/Program.cs
@@ -1,5 +1,5 @@
-using Cod3rsGrowth.Servicos;
 using Cod3rsGrowth.Infra;
+using Cod3rsGrowth.Servicos;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -7,13 +7,13 @@ ModuloInjetorServico.Servicos(builder.Services);
 ModuloInjetorInfra.Servicos(builder.Services);
 
 builder.Services.AddControllers();
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
 var app = builder.Build();
 
-// Configure the HTTP request pipeline.
+
 if (app.Environment.IsDevelopment())
 {
     app.UseSwagger();
