+++ Cod3rsGrowth.sln
@@ -9,11 +9,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cod3rsGrowth.Infra", "Cod3r
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cod3rsGrowth.Forms", "Cod3rsGrowth.Forms\Cod3rsGrowth.Forms.csproj", "{ECCCF91D-AA34-42ED-8938-6BF245E6E5FC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cod3rsGrowth.Web", "Cod3rsGrowth.Web\Cod3rsGrowth.Web.csproj", "{DB0BE585-125A-4966-9356-E3C4C39D6C0B}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cod3rsGrowth.Test", "Cod3rsGrowth.Test\Cod3rsGrowth.Test.csproj", "{96A4F732-F84E-4EDE-9AAD-95BD4D6E1219}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cod3rsGrowth.Servicos", "Cod3rsGrowth.Servicos\Cod3rsGrowth.Servicos.csproj", "{262B11EB-2CA0-4860-9BAE-7EC9C1223E3F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cod3rsGrowth.Web", "Cod3rsGrowth.Web\Cod3rsGrowth.Web.csproj", "{809D7B4F-3478-4813-A6E5-DAFD7064A0E5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,10 +33,6 @@ Global
 		{ECCCF91D-AA34-42ED-8938-6BF245E6E5FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ECCCF91D-AA34-42ED-8938-6BF245E6E5FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{ECCCF91D-AA34-42ED-8938-6BF245E6E5FC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DB0BE585-125A-4966-9356-E3C4C39D6C0B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DB0BE585-125A-4966-9356-E3C4C39D6C0B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DB0BE585-125A-4966-9356-E3C4C39D6C0B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DB0BE585-125A-4966-9356-E3C4C39D6C0B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{96A4F732-F84E-4EDE-9AAD-95BD4D6E1219}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{96A4F732-F84E-4EDE-9AAD-95BD4D6E1219}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{96A4F732-F84E-4EDE-9AAD-95BD4D6E1219}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -45,6 +41,10 @@ Global
 		{262B11EB-2CA0-4860-9BAE-7EC9C1223E3F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{262B11EB-2CA0-4860-9BAE-7EC9C1223E3F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{262B11EB-2CA0-4860-9BAE-7EC9C1223E3F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{809D7B4F-3478-4813-A6E5-DAFD7064A0E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{809D7B4F-3478-4813-A6E5-DAFD7064A0E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{809D7B4F-3478-4813-A6E5-DAFD7064A0E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{809D7B4F-3478-4813-A6E5-DAFD7064A0E5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
